Fix mobile menu aria-expanded attributes
Replaced aria-pressed with aria-expanded on mobile nav button.
Changed aria-expanded on mobile nav wrapper to data-menu-open attribute in order
to preserver the behaviour but not have the incorrect aria semantics.
Added a focus trap on mobile nav wrapper so that last focus item tabs to toggle
button and first focusable item shift tabs to toggle button.